+++ database/migrations/2020_10_04_095952_create_activity_transactions_table.php
@@ -20,10 +20,10 @@ class CreateActivityTransactionsTable extends Migration
             $table->foreign('cashier_id')->references('id')->on('users');
             $table->foreignId('activity_id')->constrained();
             $table->foreignId('category_id')->constrained();
-            $table->float('activity_price');
+            $table->integer('activity_price');
             $table->float('activity_gst');
             $table->float('activity_pst');
-            $table->float('total_price');
+            $table->integer('total_price');
             $table->boolean('returned')->default(false);
             $table->timestamps();
         });
